+++ week5/Calculator/Calculator/Extensions/String+Extension.swift
@@ -7,24 +7,6 @@
 
 import Foundation
 
-extension Formatter {
-    static let scientific: NumberFormatter = {
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .scientific
-        formatter.positiveFormat = "0.###E+0"
-        formatter.exponentSymbol = "e"
-        return formatter
-    }()
-}
-
-extension Numeric {
-    var exponentialNotation: String {
-        return Formatter.scientific.string(for: self) ?? ""
-    }
-}
-
-// MARK: insertComma
-
 extension String {
     var insertComma: String {
         let numberFormatter = NumberFormatter();
